Rename modifiers.ts to modifier-map.ts
